SetupWizard: Update wizard script to fix flow
The flow in google's suw for nougat changed a while back, which caused
the wifi network selection step to be skipped when picking the
"Copy your data" option. The workaround was to go back a step and
select the "Setup as new" option, then you could setup wifi, then go
back a couple steps to do the "Copy your data" option.

This change fixes that bug by editing the wizard script in lineage suw
to match the flow in the wizard script in google suw.

Change-Id: Ic9c9e3451af83e099e939f0ec505d24eb76c8457